pyproject.toml
[tool.poetry]
name = "hmmm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6.9"
"discord.py" = "^1.2"
asyncpg = "^0.19.0"
jishaku = "^1.16"
ujson = "^1.35"
python-dotenv = "^0.10.3"

[tool.poetry.scripts]
dev = "hmmm.bot:main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
